requirements.txt
numpy
pandas
scikit-learn
pymongo
python-dotenv

#-e .